+++ CHANGELOG.md
@@ -7,9 +7,17 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
 - Added MariaDB and SQLite to the test matrix.
+- `part_of_a_transaction` now raises an error if unhandled callbacks are detected when it starts.
+  This makes it more similar to `transaction`.
+  The error can be silenced by setting the `SUBATOMIC_CATCH_UNHANDLED_AFTER_COMMIT_CALLBACKS_IN_TESTS` setting to `False`
+
+### Fixed
 - Disallowed nesting of `part_of_a_transaction` to prevent nonsense
   implication of nested partial transactions in tests. Fixes #150.
+- `part_of_a_transaction` now clears after-commit callbacks from the transaction before it exits.
+  This avoids polluting the test's transaction with callbacks.
 
 ## [1.0.0] - 2026-04-16